Prefix from Byte Checking
Doing checks on the raw bytes rather than with translating back and forth from strings. I think this is the way I want to actually save the data when in memory and using keys and such, logging out to the user should be prettier though, rather than being in the byte format.