ActorPrefix Pretty Logging
This is kind of a proof of concept still, not sure how I want to log these out, and how I'd want to implement that. I think I'm still going to try moving fast with it though, trying out some different things.

As mentioned previously, I'd only want to use this formatting for logging out to the user, not for manipulating the data itself.

The ability to check raw bytes like this is awesome though, really opens my eyes as to how it might be really easy to write a Bedrock editing backend here in Rust, because you don't need to do string manipulation, which makes things a bit more flimsy I feel.