NBT Error Catching
Now it will skip the entity if an error occurred in the parsing stage of NBT.

Hadn't thought about it like this before, Rust does have "try/catch" statements in a way, because it makes you have to handle them because you can't actually unwrap them unless you handle the possible error on the returned value type too. I know it's not like a regular try/catch, but conceptually speaking yes it does allow you to catch errors rather than needing to work around them. And it moves the error line back a bit, which is good. In JS, you catch the error after it happens, while in Rust you handle it before it *could* happen.