Write Back to the NBT!
Writing back to the database to fix the dead entities, bringing them back to life! Turns out it actually makes a bunch of sense to allow this to make the demo world now too, so I can go test in Minecraft itself to see if this works.

Looked into how to update an existing line in the CLI (didn't end up using it for this project though). This is awesome! Probably could do it in JS too (definitely can, just curious about it).

```
// https://www.reddit.com/r/rust/comments/lhn0i2/update_the_line/

for i in 0..1000000000000000i64 {
    print!("\r{} ", i);
    std::io::stdout().flush()?;
}
```