Unliving Demo
Okay, this is a custom version of the same world, where the `Dead` state of all entities is set to `true`. I modified this code itself to *break* the key rather than fix it.

To add the custom values, I flipped the insert rule to `1`, and I made the `dead` check statically set to `true`. Then you run `--revive` funnily enough, and it will make all of the entities unliving.

Finishing up some sequencing for my album HARRY, out now on Bandcamp! It's a wild one, I'm loving it now, it's a bit more cohesive to some extent. It's meant to be a bit hair-brained though, but the collection and ordering of songs is really important to me. A song can be great, but not in the context of an album. I tried to pick only songs I enjoy listening to, and now the goal is to make the album fun too.