Automatic World Backup (Setup)
Going to add it so that the world backup will be made automatically by the program. Going to make an `.mcworld` with the current timestamp in the same folder that the world is stored in.

Didn't realize how much I really love "Launching Satellites", it's a great song! I think Marco really likes things space-related, cause he likes the SpaceX launches too; very cool :-)

https://www.discogs.com/release/3591872-Marco-Minnemann-Evil-Smiles-Of-Beauty-Sound-Of-Crime