Zip Unwrap Removals
Further trying to prevent panics.

I need to brush up on when panics are okay, and when you should add error handling to them. I might be overdoing it, but I don't know yet. I need to learn what kinds of things are meant to be what. It's still above my head a bit at the moment.